Refactor: Refactor API definition
